basic todo app with proper login and logout complted, will do refactoring to code now
