+++ UdemyCoucesPractice/Fullstack/Frontend/todo-app/src/components/Todo/ErrorComponent.jsx
@@ -1,0 +1,12 @@
+function ErrorComponent() {
+  return (
+    <div className="ErrorComponent">
+      <h1>We are really working hard</h1>
+      <div>
+        Apology for 404: reach out to our team <mark>learnFromScratch</mark>
+      </div>
+    </div>
+  );
+}
+
+export default ErrorComponent;
