+++ UdemyCoucesPractice/Fullstack/Frontend/todo-app/src/components/Todo/FooterComponent.jsx
@@ -1,0 +1,9 @@
+function FooterComponent() {
+  return (
+    <footer className="footer">
+      <div className="container">Your Footer</div>
+    </footer>
+  );
+}
+
+export default FooterComponent;
