+++ UdemyCoucesPractice/Fullstack/Frontend/todo-app/src/components/Todo/LogoutComponent.jsx
@@ -1,0 +1,10 @@
+function LogoutComponent() {
+  return (
+    <div className="LogoutComponent">
+      <h1>You are logged out !</h1>
+      <div>Thank you for visiting our website. come back soon.</div>
+    </div>
+  );
+}
+
+export default LogoutComponent;
