+++ UdemyCoucesPractice/Fullstack/Frontend/todo-app/src/components/Todo/todoApp.css
@@ -4,3 +4,10 @@ button {
   margin: 10px;
   padding: 5px;
 }
+
+footer {
+  bottom: 0;
+  position: absolute;
+  width: 90%;
+  height: 40px;
+}
